LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.canvas.fillStyle.redirect-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.pattern.canvas.fillStyle.redirect
Setting fillStyle to a pattern of an unclean canvas makes the canvas origin-unclean
Actual output:

PASS Setting fillStyle to a pattern of an unclean canvas makes the canvas origin-unclean
